Use a mutex to avoid race conditions corrupting browsers.json
The specific race here is only hit when no browers.json exists.
In that case, JBL attempts to detect & write the browser config
twice, in parallel. If that happens, and the config doesn't
have exactly the same length in both cases, the write ends up
with spurious data appended, and the config file no longer
parses. We can recover from that, but it breaks in the meantime.
With this change, it should (in theory) never happen.