Ensure PHP interception works on OSX
On OSX, we can't prepend to PATH, because a script called path_helpers
prepends all standard paths to the PATH in every new shell, effectively
overwriting us. To fix this, we edit each of the common shell startup
config files, which gives us a hook into shells _after_ this happens.
The hooks are behind an env var flag (HTTP_TOOLKIT_ACTIVE), so have
no effect the rest of the time, and are cleared up when the terminal
closes (n.b. on OSX we get proper process exit events that make this work)