Fix the 'no terminal detected' error flood
A small number of users are hitting this issue, and sending error reports
every 10 seconds while the app is open. This is annoying.

For now, I've disabled the existing Sentry DSN. This commit moves to a
new one (for the exact same project), and ensures that reports are
only ever sent once.