Return metadata from the activate/deactivate API
Previously this just returned a boolean. This is technically a breaking
change, as it now returns a JSON object, with the result in the .success
field, but nobody should be depending on this internal API except the
UI, and that doesn't depend on the result for anything substantial.

Using JSON here instead of a proper result is a compatibility
workaround. In future, the plan is to remove GraphQL entirely, which is
why I'm not extending the complexity of the schema at this stage.