Tighten the CSRF security, since Electron interceptor allows RCE
Previously the API was difficult to CSRF, with attacks that only
worked in certain old browsers etc. With this change, it should
be impossible to send a request to the GraphQL API that isn't
from an allowed server. There is still a remaining risky avenue
due to Mockttp's API (if you're clever, you could spoof a trusted
origin directly), but that'll be closed the same way shortly.