Fix issues with FF first-run setup in OSX, and disable it in Windows
Previously this only worked in Linux. In Windows, it launches a child
process, so we can never get the right pid, so we can't reliably kill it
and tidy up as we'd like, and the two-step startup *requires* this, or
the second step fails with an error.

On Mac, we had a similar problem, but actually it turns out we just
needed to wait properly for the exit to propagate and shut everything
down. There is a child process (open -a ...), but it shuts down ok
anyway.