Quick fix for test errors when killing firefox
These only occur in Node 12 & 14 (not the production node
version anyway), and they seem to be irrelevant. Yes, we
see connection resets when the browser shuts down, but
that's OK. It doesn't look like there's any cases where
these are real uncaught exceptions (not reproducible
manually, so far at least...).