Allow auth tokens to be set by $HTK_SERVER_TOKEN
This tightens token auth further. Using --token, another process on
your machine could in theory look at process command line to get
the auth token, and use that to access the API to increase its
privileges.

There's no known viable attacks even given that access, and this would
require a malicious process to already be running on your machine, but
it's good to be able to block that regardless.