Add a timeout to ADB root detection commands
It turns out that on some devices "su root whoami" does not
fail or run whoami - it just successfully starts a shell.
In these cases, we have a problem, as that blocks the rest
of the root detection. To fix this, we now time out in any
such cases. It would be a bit better to properly get root
using the interactive shell, but the practicality is a bit
complicated, so let's ignore that for now until a real
device that needs that comes up.