Intercept Ruby HTTP even when net/http isn't imported
In some cases, just net/https may be imported. This works
if net/http is imported too, but if not then interception
relies entirely on the SSL_CERT_FILE variable. Unfortunately,
in some environments (notably including Mac's default Ruby)
that variable is ignored. With this change, we ensure that
even if just net/https is required, our HTTP hook is always
enabled.