Clean up activation timeout logic, which didn't work as intended
Seems like I misunderstood/confused myself somehow in cf3ea7e3: this
doesn't configure an activation timeout (which doesn't really exist,
though we report errors at 30s+) but instead sets the isActivable
timeout. That's not actually very important for ADB (should be quick,
perhaps after the first check) AFAICT, and it's definitely not what was
described/documented for this change.

Now, we extend the time a little for the JVM and ADB, both of which may
need to wait for external processes (especially slow on Mac) but
otherwise keep the default 1s for isActivable and 30s warning for
activation itself.