Update to Node 14.17 (hopefully improves HTTP/2 behaviour)
For some reason, this breaks the npm installs that occur as part of the
build script - postinstall steps can no longer access the install
directory, and everything fails. This change also enables unsafe-perm
only in CI, so that we can npm install as root there. I can't see any
risk to this - anything that could be exposed from within the CI build
that's interesting would be in the env or build workspace anyway.