Ignore Buffer() deprecation warning
This is caused by the ftp module, used in the pac-proxy-agent module,
because PAC URLs may be FTP. That proxy agent is in turn used inside
Mockttp for proxy support. This only appears in built packages
annoyingly, because it only fires for non-node_modules deprecations.

The issue itself isn't actually a security issue, it's here:
https://github.com/mscdex/node-ftp/blob/7dff82fc049a52f51803bdffb95ec1de383f9fac/lib/connection.js#L53