Fix space escaping bug in existing-terminal
When used while HTTP Toolkit was running in a path containing a space,
the quotes in NODE_OPTIONS weren't escaped in the shell script that
included it. That made the shell script invalid, breaking it entirely.
We now ensure that all env var values have any quotes included correctly
escaped.