Add a Ruby override for Cloudinary
Cloudinary uses rest-client internally, which is intercepted
correctly, but Cloudinary itself overrides the proxy setting
and doesn't use any of the standard proxy env vars to do so.
It seems to have its own CLOUDINARY_API_PROXY env var, but
it doesn't seem to work? Easiest & cleanest to just override
the config in the gem directly instead.