Use host.docker.internal inside containers, to support Mac
This is automatically supported on Mac & Windows, so that's great. I
suspect it'll work for unusual networks too, and Docker will handle the
complexity, big win.

On Linux, we're still assuming a normal default bridge network, but we
can move to 'host-gateway' later, based on support for that, or
potentially examine the network config dynamically or similar.