Resolve relevant Docker aliases in our DNS
This is important for non-trivial Docker setups that communicate on
their own network. When they do so, they'll usually use some sort
of container alias, which Docker will resolve but normal DNS cannot.
To handle that, we monitor Docker events, track the containers we're
intercepting, and ensure all the hostnames relevant to them are
always resolveable.