Allow intercepting already created docker-compose containers
By changing the filters used by DC's container queries, we effectively
force DC to recreate all containers, which allows us to intercept them
even if there's an existing valid unintercepted container that could be
used. By renaming the created containers, we allow that to happen
without deleting the existing container, in case the user wants to use
it later. We carefully don't allow this for the case where the
container is actually running though, because otherwise we'd rapidly
run into conflicts on volumes (might cause data loss) and network
aliases (unclear what happens here, but definitely something unexpected).

In theory, we could stop the running containers automatically instead,
but that's a bit invasive, and how long you wait for shutdown before
killing them is config & context specific (e.g. to ensure DBs are all
flushed to disk). Better to leave it up to the user.