Revert "Use a custom hostname for Docker interception"
This reverts commit 08020b037c60a933a0cbe2461dec2fa5dfb4d133.

This breaks interception on Windows (and presumably Mac) because
host.docker.internal is used as the target IP for the host mapping, and
that's not an IP. We could use host-gateway, but that's not widely
supported yet (Engine 20.10/API 1.41 only). Only a minor benefit, so
just rolling this back for now, so we can use host.docker.internal
everywhere and get maximum easy compatibility in the short term.