WIP: [skip ci] Reintegrate DNS server, for edge case tunnel resolution
Without this, we had to resort to very slow and messy hacks like
rewriting /etc/hosts in the tunnel using 'docker exec' and bash
utilities. That's madness. Making Mockttp use our DNS so we have
complete control over resolution is much better (and much easier to
debug if we get odd DNS results later on).