Expose a shutdown API to support clean shutdown on Windows
Windows can't send signals like SIGINT/SIGTERM to arbitrary processes in
a way that can support clean shutdown. With Docker, we have quite a bit
of cleanup to do, and clean shutdown becomes more important, so it's
useful to expose an API that the desktop app can use to trigger this
directly.