Fix pipe issues in Windows docker-compose interception
Windows named pipes aren't handled well by Python (seemingly) when
used in docker-compose (<v2, which isn't stable yet). When they are
closed cleanly (at least from Node's POV...) it throws an error,
either complaining that it's closed on read or on write (which gets
reported as an odd "incompatible API" error incorrectly, see
https://github.com/kenneth-git/compose/blob/d06e15323a44b3a544969cb77b517d0dac295a94/compose/cli/errors.py#L67-L72
for background).

On the other hand, it does properly close down attach connections
all by itself, which works around the need for actively closing
the connection on our end, so we can just disable that for this
specific case and everything seems to work.