Dramatically improve docker-compose interception UX
This moves from remapping the project name label to instead remapping
the labels that hold the config hash. That means that all docker-compose
commands (intercepted or not) are now working on the same state, with
the only difference being that each one considers containers created by
the other to need recreating. This is great! No more conflicts, no
leftover container usage, and much easier container management, with a
simpler implementation. What a win!