Use npm 8 in CI (to allow overriding dep versions)
This requires some fixes, including:
- Updates to support the new base image, which otherwise has some issues
  with npm permissions in this case.
- Tweaks to the release build process to ensure that the prepare script
  keeps working correctly with the different lifecycle behaviour.
- Pinning prebuild-install to v7, because v5 (used by some deps) doesn't
  support the env vars for cross-platform builds with npm v7+.

Regarding the base image: previously, the image had a default root
user, so the checkout was made as root, but GHA always runs run steps
as user 1001. After updating npm, it started to fail to install due to
permissions issues, because it was running as user 1001 while the
checkout was made as root.

The base image has now been updated to default to a 1001 user called
'build-user' who has passwordless sudo & docker permissions. With a
few small changes here, we can support that and everything should work
correctly with a single consistent user used in every step.