Revert "Switch from @oclif/dev-cli to Oclif"
This reverts commit 63dcb54c05d7affb33598d8119be19a7bfcc30dd.

This is due to the new Oclif package not supporting templated file names
& paths, which makes it impossible to easily deploy to GitHub Releases.
Once templates are supported, we can fix this backup. Just need to make
sure that the manifest file names match, update ci.yml to correctly
upload them, and make sure that the paths inside the manifest are
correct.