Fix recreation of aliases/network config in Docker container attach
Previously we didn't touch 'NetworkMode' which meant that it retained
the name of the main network the container was using, and that meant
that it reconnected to that automatically, ignoring the subsequent
network connection call we made with the full settings.

We now connect all attached containers to 'none' initially, and then
manually reconnect them to the original networks with their full
settings, to preserve everything (especially aliases).

This change also ensures we wait for network setup to be complete, so
the container starts with this change entirely in place.