Fix Chrome flags writing so it works on more types of rooted devices
Previously it assumed that 'adb push' would work for root-only paths,
which wasn't true for devices that expose an explicit 'su' binary we can
use, where we don't use 'adb root'. We now push a script that does the
writing and then run it with root directly, which should work for all
root setups.