Improve JS interception when using ESM imports (e.g. node-fetch)
This is not a proper fix. We don't yet use loaded hooks, because they're
experimental, apparently _extremely_ likely to change in future, and
setting the option would cause issues with older node versions.

We can look at this in future once the API has stabilised (by doing
version detection in the PATH wrapper) but for now we can solve the
immediate issue by just enabling global-agent 100% of the time
regardless, which avoids issues with hooks not firing for ESM
node-fetch. I think ESM usage in Node is still not widespread, so this
should be enough to keep things usable in the short term.