Hook hgdemandimport to fix Mercurial in intercepted terminals
Without this, 'hg' fails with:

ModuleNotFoundError: No module named 'http.cookiejar

Not totally clear why, but it's fairly clearly an interaction between
our funky import hooks and their funky import hooks. This module
appears to be a startup time optimization, and it's certainly optional
(there's an env var to disable it, and a built-in list of modules it
doesn't support) so best to just no-op it's enable() function.