Don't inject DOCKER_HOST proxy address into Docker containers
This could work in theory. We'd have to inject the variable along with
mounting the socket itself though. We'd only want to do that in cases
where we were sure that the socket was already being mounted (or its a
big security risk), and that's a) hard to detect and b) extra hard to
redirect. Doable, probably, but not worth it for now.

It's also possible that some DinD setups might work regardless. If they
check DOCKER_HOST to find the path to the socket, they'll end up
mounting our intercepted socket correctly all by themselves.