Be more flexible in Electron launch paths for .app bundles
Previously, we needed the exact executable path, or the app path. Now we
also accept the app base name (/Applications/Slack) even if the real
path is .app, because that extension is hidden on Mac. This is useful
because we're going to have to move to manually entering app paths on
Mac, since in recent Electron versions the file picker no longer works
for selecting inside .app bundles.