Skip cpu-features when bundling - causes problems & not necessary
Specifically, this seems to cause issues when building in some
environments where the build dependencies are all available (and so the
build succeeds) but then bundling fails because it's native. We never
need this, honestly we don't really need ssh2, it's probably not usable,
and it's optional anyway, so it's best to just drop it entirely.