Fix WebRTC bug in production installations only, due to config perms
When running locally, or running in an updated server environment, the
webextension folder, including its config, is user writable.
Unfortunately, when running in a fresh production install, that's not
true: the folder is usually administrator/root-owned, and so cannot be
modified by users, so our config injection mechanism didn't work.

To fix this, we now 'install' the webextension on first use, copying it
(~10 files total, 1MB) into a temp directory, which can then be
modified, and deleted again when the app exits.