Improve handling of Docker host-gateway routing
Previously, we explicitly handled the host.docker.internal alias,
treating it as non-routeable (so that it was handled locally, and the
mapping to 127.0.0.1 went to the host). This was tied to the specific
alias, and didn't support other EXTRA_HOSTS mappings. This is now
expanded, with custom mappings and other standard aliases also
included.

This also reduces our dependency on host.docker.internal generally,
which we're going to move away from for proxy config imminently.