Refactor Docker proxy host routing to avoid ExtraHosts
ExtraHosts doesn't work with any client that relies exclusively on DNS
(one good example so far - Mastodon - but this is also easy to do in
Node.js and elsewhere). This therefore breaks all Docker connections to
the proxy on Linux, where host.docker.internal was injected this way.
Instead, we now use the raw IP on Linux, avoiding DNS/hostname issues
entirely.