Inject content into Docker via volumes, not bind mounts
This is useful, because a) bind mounts are slow and problematic (they
come with various warnings) on Docker Desktop on Windows & Mac) and b)
this moves us slowly towards a world where we can successfully intercept
a _remote_ Docker host (though we still can't yet, mostly because we
don't know the right proxy address to use for HTTP Toolkit).

Note that unlike other Docker services, we ensure the Docker volume is
running synchronously on-demand, not just async when we see Docker
proxy activity.

Without that, if the volume didn't exist for some reason (shouldn't
happen, but possible given races etc or deletion elsewhere) then
containers could've been created referencing it when it didn't exist,
whereupon Docker would create it automatically and break everything.