Fix Docker tunnel pull race condition
Previously we assumed that when pull() resolved it was complete. This is
false. In reality we have to wait for the resulting stream to explicitly
finish, so we had a subtle race condition here where the tunnel would
not be available if it was required immediately after startup (e.g. in
the tests) on the first run of the app, on a weak connection or whilst
Docker hub was running a little slow.