Reduce requests required to check that Docker tunnel is running
Previously we checked on the image and then always checked on the
container too. Now, in the very common case where the container is
already running, we just check that and call it a day (we should never
need to pull the image if the container is already running anyway).