Fix rule parameter query API to fix Docker tunnelling on Mac/Windows
Previously Docker on Mac/Windows tunneling between containers generally
didn't work, because the UI did not know that the rule parameter to
proxy via the tunnel was available, as it wasn't updating in the API.

The API now always returns the latest value for this. This is risky for
race conditions, but that's difficult to avoid and I think it should be
safe: session starts, Docker setup sychronously sets the rule parameter
for the new session, while the UI must wait async for the start to be
confirmed before it can request the corresponding config (which requires
the port, so can't be parallelized) and read the param keys.