Fix Windows $PATH overrides in powershell, and simplify en route
We didn't use this var previously since it wasn't set until the
recent-ish PHP scandir changes, but it's definitely better to just pass
the path explicitly rather than trying to detect it.

Haven't changed Linux just to avoid churn because it's a little
complicated to confirm the equivalent and test every possible case, but
we absolutely could if we ever have issues with the current setup there.