Test distributables against multiple runner versions, not just latest
This caused the release issues with v1.12.4, because we only tested
against Mac v12, and node-datachannel had issues on v11. We were testing
v11 in the past, but macos-latest updated in October 2022.

We include MacOS 10.15 here just for now, but note that it will start to
hit brownouts in July 2023, and at that point we should drop it
entirely.

This also updates the OS versions used in build & publish CI steps too.
Oddly it seems GHA hits an internal error before starting to build if
the (pre-test) build step OS version isn't also updated? Very odd.