Don't include IPv4 or Docker addresses in connection config sent via ADB
This fixes a bug where devices would connect via IPv6, which isn't
actually supported. This worked with QR codes (because the UI does
correctly do this filtering) but not with ADB because of this issue. It
also reduces unnecessary network checks (and waiting for timeouts,
potentially) for some cases too.