Fix ADB root detection command quoting & add another possibility
Quoting was definitely wrong here - it turns out that ADB is actually
extra-quoting our arguments under the hood, so we were effectively doing
it twice, and this definitely didn't work as intended. The new logic
definitely does send the correct command however, so should work for the
devices affected by that issue.

This also adds a similar case for the 'su root' devices, just in case
some are also affected, and reorders the commands a little (I think this
order is preferably, but really it shouldn't make much difference given
the new unambiguous testing approach).