Fix Electron interception issue
On some systems, with the new Node, CDP connections default to ::1
instead of 127.0.0.1, and it seems the Electron debug port is only
accessible on 127.0.0.1 by default, so this doesn't work.

We now fix both the listening address & target address explicitly, which
should ensure these reliably match up correctly in future.