Update Mockttp for relative-localhost, MacOS+OpenSSL & Node 20 fixes
Nothing super impactful here, but all good-to-do changes. Both the
Mac OpenSSL issue & Node 20 fixes specifically don't do anything yet,
but will help to put us in a better place to upgrade to newer Node
releases later this year, once other issues there are resolved.

Relative localhost refers to handling of requests to 'localhost'
addresses from non-local clients (which should be routed back to
themselves, not to our own localhost). Previously this was applied
textually, now we fully resolve the target IP and check definitively
whether it's a localhost address or not.