Update Node for perf boosts, header removal fixes, and to stay on LTS
v16 is now EOL so we need to upgrade pretty pronto. All issues are now
resolved as far as I'm aware, so this should work correctly with H2
(previously there were issues there in early v20) and with ALPN once
Mockttp is updated (previously unrecognized ALPN was now rejected, which
could cause problems).