Update very outdated 'engines' package.json field
This isn't really useful in the traditional way (this project isn't
intended to be used as Node package directly) but it's used by Oclif and
it's good to be explicit anyway. In general, this should be ^ of the
Oclif node version from package.json, and will change unpredictably to
match that in future as that changes. This isn't a breaking change,
because the node version used is supposed to be the one from
package.json anyway - BYON will work, but it's not officially supported
or worried about.