Disable chromium component updates without disabling the components
This is arguably a Chromium bug really, but c'est la vie. When disabling
updating explicitly, the components are disabled completely (nothing is
shown on chrome://components) which causes problems and unrepresentative
behaviour, most notably when testing services that use DRM video.

This replaces that with an invalid update URL, so the update fails
instead. That results in a single failing request in HTTP Toolkit, but
this is better than either broken functionality or full updates (many
large requests every time).