Fix up the test environment setup script
The real issue here is that this depends on our build base image, which
is based on ACT's base scripts (no longer maintained) and which doesn't
fully work any more due to various outdated assumptions (particularly
around things like installing Chromium). We're going to have to deal
with this eventually, but not today!

Hopefully eventually we can do so by migrating to something based on
the normal GHA images (but a bit more stability and versioning would be
nice...) but who knows.