Fix issues with Opera startup
Opera seems to occasionally do a quick start-restart flicker as it
loads in recent versions, resulting in it reloading the hide-warning
page, which is no longer available after the first request.

No reason for us to stop that server quite so aggressively, so we now
keep it up for 10 seconds just in case this happens so more than one
load will work just fine.