Fix issue that could insta-crash Opera in some launches
It looks like new Opera releases have a UI bug where opening a URL that
closes while the tab animation is happening (i.e. instantly) can crash
the browser. It reloads and recovers (which is why the previous commit
was required) but it would be nice to not do that. If we delay just 1ms
this doesn't seem to happen any more, so let's do that.