Try to fix BackgroundServiceStartNotAllowedException for Android stop
We see quite a few of these exceptions. We're allowed to send the intent
to the Android app, but once it receives it, it's still in the
background, and it's not allowed to send an intent to the background VPN
service, so the service never actually shuts down. Doesn't seem to
happen all the time, but just on some devices. This fix should clean
that up nicely.