Update TS target version
Should be very safe, since ES2020 supports back to Node 14, and we're
running far ahead of that in the prod node version at least. Main goal
is just avoiding syntactic desugaring, which adds lots of noise (and
probably perf hits) for things like async/await that we'd like to avoid.