Drop MacOS 11 from automated builds due to GitHub removal
This won't cause any immediate changes for users, but it does
dramatically increase the chance that we'll miss an incompatibility
(e.g. in some native dependency like node-datachannel) that no longer
works on Mac 11. Unfortunately there's not much we can do about this -
GitHub just started brownouts for MacOS 11 today and it'll be totally
gone in two weeks.