Add support for skipping known-problematic ports in Frida native capture
This only affects Frida's native connect hook, which forcibly captures
traffic that ignores the proxy server. In most cases that's fine and
important (e.g. for Flutter apps) but in certain scenarios where the
traffic is completely non-HTTP it can cause problems. We could make this
configurable later, but for now we just track a few high profile cases.