Fix Android hash subject calculation
This broke Android setup with new certificates, because changes to the
certificate generation code resulted in a subject starting with 0, which
was lost in the filename, and so never matched. This bug has always been
present, but wasn't visible until that took effect.

This now left-pads with 0s to resolve this issue.