Fix first-run Frida interception
On Android (at least) when an app has never been launched,
Socket.connect() never resolves until the app is resumed. In our case,
we need to pick an IP & inject the hooks before we can resume, so this
is problematic.

To handle this, we already fallback to the tunnel IP if it connected OK
(should almost always happen) but the 10s timeout means the connection
actually breaks before that can happen. Dropping the timeout (connect
should really take single-digit milliseconds in normal cases) means we
inject the script & resume very quickly in this case.